requirements.txt
comtypes==1.4.13
dnspython==2.8.0
numpy==1.26.4
opencv-contrib-python==4.5.5.64
pandas==2.3.3
pillow==11.3.0
pymongo==4.15.3
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pywin32==311
six==1.17.0
tzdata==2025.2
